Enable highlighting for css.resolve/css.global
Prior to this commit, css on it's own would correctly syntax highlight everything between it's backticks, but not css.resolve or css.global. I think this was due to a TextMate regex quirk -- anything.resolve *should* have matched, but it was looking for the start of a word boundary.

The fix is to add `\\bcss` before checking for the `.resolve` or `.global` bits. Sure, this is a bit more restrictive (you can't have anything.resolve anymore), but css is already hard coded for the css`` case.