fix: minor styling issue on last pattern
On version 1.5.0, case `<style global jsx>` had broken tag highlights. On version 2.0.0, case `<style jsx global>` has broken tag highlights as a result of it being switched as the last OR case. Not entirely sure what the cause of this might be, but there is no need to overcomplicate the expression cases and make `jsx`, `jsx global`, and `global jsx` as a single capture group.